+++ src/oci/load_balancer/models/ssl_configuration_details.py
@@ -38,6 +38,10 @@ class SSLConfigurationDetails(object):
             The value to assign to the verify_peer_certificate property of this SSLConfigurationDetails.
         :type verify_peer_certificate: bool
 
+        :param has_session_resumption:
+            The value to assign to the has_session_resumption property of this SSLConfigurationDetails.
+        :type has_session_resumption: bool
+
         :param trusted_certificate_authority_ids:
             The value to assign to the trusted_certificate_authority_ids property of this SSLConfigurationDetails.
         :type trusted_certificate_authority_ids: list[str]
@@ -67,6 +71,7 @@ class SSLConfigurationDetails(object):
         self.swagger_types = {
             'verify_depth': 'int',
             'verify_peer_certificate': 'bool',
+            'has_session_resumption': 'bool',
             'trusted_certificate_authority_ids': 'list[str]',
             'certificate_ids': 'list[str]',
             'certificate_name': 'str',
@@ -78,6 +83,7 @@ class SSLConfigurationDetails(object):
         self.attribute_map = {
             'verify_depth': 'verifyDepth',
             'verify_peer_certificate': 'verifyPeerCertificate',
+            'has_session_resumption': 'hasSessionResumption',
             'trusted_certificate_authority_ids': 'trustedCertificateAuthorityIds',
             'certificate_ids': 'certificateIds',
             'certificate_name': 'certificateName',
@@ -88,6 +94,7 @@ class SSLConfigurationDetails(object):
 
         self._verify_depth = None
         self._verify_peer_certificate = None
+        self._has_session_resumption = None
         self._trusted_certificate_authority_ids = None
         self._certificate_ids = None
         self._certificate_name = None
@@ -150,6 +157,40 @@ class SSLConfigurationDetails(object):
         :type: bool
         """
         self._verify_peer_certificate = verify_peer_certificate
+
+    @property
+    def has_session_resumption(self):
+        """
+        Gets the has_session_resumption of this SSLConfigurationDetails.
+        Whether the load balancer listener should resume an encrypted session by reusing the cryptographic parameters of a previous TLS session, without having to perform a full handshake again.
+        If \"true\", the service resumes the previous TLS encrypted session.
+        If \"false\", the service starts a new TLS encrypted session.
+        Enabling session resumption improves performance but provides a lower level of security. Disabling session resumption improves security but reduces performance.
+
+        Example: `true`
+
+
+        :return: The has_session_resumption of this SSLConfigurationDetails.
+        :rtype: bool
+        """
+        return self._has_session_resumption
+
+    @has_session_resumption.setter
+    def has_session_resumption(self, has_session_resumption):
+        """
+        Sets the has_session_resumption of this SSLConfigurationDetails.
+        Whether the load balancer listener should resume an encrypted session by reusing the cryptographic parameters of a previous TLS session, without having to perform a full handshake again.
+        If \"true\", the service resumes the previous TLS encrypted session.
+        If \"false\", the service starts a new TLS encrypted session.
+        Enabling session resumption improves performance but provides a lower level of security. Disabling session resumption improves security but reduces performance.
+
+        Example: `true`
+
+
+        :param has_session_resumption: The has_session_resumption of this SSLConfigurationDetails.
+        :type: bool
+        """
+        self._has_session_resumption = has_session_resumption
 
     @property
     def trusted_certificate_authority_ids(self):