+++ CHANGELOG.rst
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 
 The format is based on `Keep a Changelog <http://keepachangelog.com/>`_.
 ====================
+2.126.3 - 2024-05-14
+====================
+
+Added
+-----
+* Support for the Marketplace Private Offer service
+* Support for resizing local file systems in the Database service
+* Support for PPv2 (Proxy Protocol version 2) in the Load Balancer service
+* Support for custom return path for sent emails in the Email Delivery service
+* Support for session resumption in the Load Balancer service
+ 
+====================
 2.126.2 - 2024-05-07
 ====================